[ci skip] [WIP] Rebased and fixing builds

+++ src/jet/point_parallel_hash_grid_searcher2.cpp
@@ -333,7 +333,7 @@ void PointParallelHashGridSearcher2::deserialize(
 
     // Copy simple data
     auto res = fbsToJet(*fbsSearcher->resolution());
-    _resolution = {res.x, res.y};
+    _resolution = {(ssize_t)res.x, (ssize_t)res.y};
     _gridSpacing = fbsSearcher->gridSpacing();
 
     // Copy points
